move utils to nlp/huggingface
